backend/requirements.txt
Django>=5.1.3,<5.2
psycopg2>=2.9.10,<2.10
djangorestframework>=3.15.2,<3.16
djangorestframework_simplejwt>=5.3.1,<5.4
pytest-django>=4.9.0,<4.10
drf-spectacular>=0.27.2,<0.28
